\name{ProjectiveDecomposition-package}
\alias{ProjectiveDecomposition-package}
\alias{ProjectiveDecomposition}
\alias{pd}
\alias{pd.default}
\alias{pd.dgTMatrix}
\docType{package}
\title{
\packageTitle{ProjectiveDecomposition}
}
\description{
\packageDescription{ProjectiveDecomposition}
}
\usage{
pd(A,)
\method{pd}{default}()
\method{pd}{dgTMatrix}()
\method{summary}{ProjectiveDecomposition}()
\method{sim}{ProjectiveDecomposition}(pd,M)
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ProjectiveDecomposition}
\packageIndices{ProjectiveDecomposition}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ProjectiveDecomposition}

Maintainer: \packageMaintainer{ProjectiveDecomposition}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
